Move formatter to its own file
